requirements.txt
click>=8.0.0
pygit2>=1.12.0 # Check latest compatible version
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.25.1
numpy>=1.21.0
# For testing (optional for basic run, but good to have)
pytest>=6.2.4